third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/source-frame-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that it's possible to set breakpoint in source frame, and that source frame displays breakpoints and console errors.


Running: testConsoleMessage
Script source was shown.
Message added to source frame: test error message
Message removed from source frame: test error message

Running: testShowResource
